+++ src/components/dropdown-filter/dropdown-filter.js
@@ -24,6 +24,9 @@ import escapeStringRegexp from 'escape-string-regexp';
  * You can also use the component in 'suggest' mode, which only shows the dropdown
  * once a filter term has been entered.
  *
+ * You can also use the component in 'freetext' mode, which behaves like 'suggest',
+ * but allows write-in text values in addition to list options.
+ *
  * You can also define a function using the 'create' prop, this will allow you
  * to trigger events to create new items.
  *
@@ -31,7 +34,6 @@ import escapeStringRegexp from 'escape-string-regexp';
  * @constructor
  */
 class DropdownFilter extends Dropdown {
-
   constructor(...args) {
     super(...args);
 
@@ -92,7 +94,15 @@ class DropdownFilter extends Dropdown {
      * @property suggest
      * @type {Boolean}
      */
-    suggest: React.PropTypes.bool
+    suggest: React.PropTypes.bool,
+
+    /**
+     * Should the dropdown accept free text as well as suggested options?
+     *
+     * @property freetext
+     * @type {Boolean}
+     */
+    freetext: React.PropTypes.bool
   }
 
   /**
@@ -116,8 +126,9 @@ class DropdownFilter extends Dropdown {
    * @param {String} val
    */
   selectValue(val, visibleVal) {
+    let filter = this.props.freetext ? visibleVal : null;
     super.selectValue(val, visibleVal);
-    this.setState({ filter: null });
+    this.setState({ filter: filter });
   }
 
   /*
@@ -132,7 +143,7 @@ class DropdownFilter extends Dropdown {
       highlighted: null
     };
 
-    if (this.props.suggest && ev.target.value.length <= 0) {
+    if ((this.props.suggest || this.props.freetext) && ev.target.value.length <= 0) {
       state.open = false;
     } else {
       state.open = true;
@@ -155,8 +166,17 @@ class DropdownFilter extends Dropdown {
    */
   handleBlur = () => {
     if (!this.blockBlur) {
-      let filter = this.props.create ? this.state.filter : null;
+      let filter = null;
+
+      if (this.props.create || this.props.freetext) filter = this.state.filter;
       this.setState({ open: false, filter: filter });
+
+      if (this.props.freetext) {
+        let opt = this.props.options.find((opt) => { return opt.get('name') === this.state.filter; });
+
+        if (opt) this.selectValue(opt.get('id'), opt.get('name'));
+        else this.emitOnChangeCallback('', this.state.filter);
+      }
     }
   }
 
@@ -166,7 +186,7 @@ class DropdownFilter extends Dropdown {
    * @method handleFocus
    */
   handleFocus = () => {
-    if (!this.props.suggest && !this.blockFocus) {
+    if (!(this.props.suggest || this.props.freetext) && !this.blockFocus) {
       this.setState({ open: true });
     } else {
       this.blockFocus = false;
@@ -192,7 +212,7 @@ class DropdownFilter extends Dropdown {
    * @param {Object} options Immutable map of list options
    */
   prepareList = (options) => {
-    if ((this.props.suggest || !this.openingList) && typeof this.state.filter === 'string') {
+    if ((this.props.suggest || this.props.freetext || !this.openingList) && typeof this.state.filter === 'string') {
       let filter = this.state.filter;
       let regex = new RegExp(escapeStringRegexp(filter), 'i');
 
@@ -280,6 +300,17 @@ class DropdownFilter extends Dropdown {
   }
 
   /**
+   * Extends the input content to include the input icon.
+   *
+   * @method additionalInputContent
+   */
+  get additionalInputContent() {
+    let content = super.additionalInputContent;
+    if (this.props.suggest || this.props.freetext) content.shift();
+    return content;
+  }
+
+  /**
    * Returns the list options in the correct format
    *
    * @method options
@@ -306,12 +337,8 @@ class DropdownFilter extends Dropdown {
    * @method inputClasses
    */
   get inputClasses() {
-    return classNames(
-      super.inputClasses,
-      {
-        'carbon-dropdown__input--filtered': !this.props.create && typeof this.state.filter === 'string'
-      }
-    );
+    let filtered = !this.props.create && !this.props.freetext && typeof this.state.filter === 'string';
+    return classNames( super.inputClasses, { 'carbon-dropdown__input--filtered': filtered });
   }
 
   /**
@@ -373,7 +400,6 @@ class DropdownFilter extends Dropdown {
 
     return newValue;
   }
-
 }
 
 export default DropdownFilter;